Added Admin Auth, Department & Default Leaves API Model and Component

+++ UI/src/app/HR/data-entry/data-entry.component.ts
@@ -14,6 +14,8 @@ import 'fullcalendar';
 import { Router } from '@angular/router';
 import { UserDataService } from 'src/app/services/user-data.service';
 import { DatePipe } from '@angular/common';
+import { SettingsService } from 'src/app/services/settings.service';
+import { p } from '@angular/core/src/render3';
 //declare var $: any;
 
 @Component({
@@ -35,6 +37,7 @@ export class DataEntryComponent implements OnInit {
   editEmpFlag = false;
   editHolidayFlag = false;
   holidayList = [];
+  departmentList = [];
 
 
   @Input()
@@ -47,7 +50,7 @@ export class DataEntryComponent implements OnInit {
 
   defaultConfigurations: any;
   constructor(private router: Router, private userLoginService: UserLoginService, private userDataService: UserDataService, private datepipe: DatePipe,
-    private holidayService: HolidayService) {
+    private holidayService: HolidayService, private settingsService: SettingsService) {
     this.user = new UserModel()
     this.holiday = new HolidayModel()
 
@@ -141,6 +144,20 @@ export class DataEntryComponent implements OnInit {
       }, (error) => {
         console.log(error);
       })
+      this.settingsService.settingsData().subscribe((response) => {
+        this.departmentList = JSON.parse(response["_body"]).departmentList;
+        for (let i = 0; i < this.employeeList.length; i++) {
+          var deptId = this.employeeList[i].department;
+          var departmentArray = this.departmentList.find(p => p._id === deptId);
+          if (departmentArray) {
+            this.employeeList[i].departmentName = departmentArray.departmentName
+          }
+        }
+
+        console.log(this.departmentList)
+      }, (error) => {
+        console.log(error);
+      })
     }
     else {
       this.logAdmin = false;
@@ -165,7 +182,6 @@ export class DataEntryComponent implements OnInit {
     this.user.employeeType = '';
     this.user.lastName = '';
     this.user.email = '';
-    this.department = ['Administration', 'Business Systems', 'CAD/CAM', 'IT', 'Marketing', 'Services', 'Support']
     //this.addEmployeeForm.resetForm();
   }
 
@@ -197,13 +213,14 @@ export class DataEntryComponent implements OnInit {
     }
   }
 
-  addEmployee() {
+  addEmployee(resetForm) {
     this.checkDuplicateEmpEmail();
     if (localStorage.getItem('adminToken')) {
       if (!this.isEmployeeCodeExist && !this.isEmployeeEmailExist) {
         this.userDataService.adminAddEmployeeData(this.user).subscribe((response) => {
           this.newUserData = JSON.parse(response["_body"]).user;
           alert('Employee added')
+          this.user = new UserModel();
           this.onloadList()
         }, (error) => {
           console.log(error)
@@ -216,6 +233,7 @@ export class DataEntryComponent implements OnInit {
         this.userDataService.addEmployeeData(this.user).subscribe((response) => {
           this.newUserData = JSON.parse(response["_body"]).user;
           alert('Employee added')
+          this.user = new UserModel();
           this.onloadList()
         }, (error) => {
           console.log(error)
@@ -225,65 +243,128 @@ export class DataEntryComponent implements OnInit {
     }
   }
 
-  editEmployee(editEmployee) {
-    debugger;
+  editEmployee(editedUser) {
     this.editEmpFlag = true;
-    this.user = editEmployee;
-    this.user.dateOfJoining = this.datepipe.transform(this.user.dateOfJoining, "yyyy-MM-dd");
-    this.user.resignationDate = this.datepipe.transform(this.user.resignationDate, "yyyy-MM-dd");
-    this.user.leavingDate = this.datepipe.transform(this.user.leavingDate, "yyyy-MM-dd");
+    if (editedUser.password) {
+      this.user.password = editedUser.password;
+    }
+    debugger;
+    this.user.isHR = editedUser.isHR;
+    this.user.employeeCode = editedUser.employeeCode;
+    this.user.firstName = editedUser.firstName;
+    this.user.lastName = editedUser.lastName;
+    this.user.employeeType = editedUser.employeeType;
+    this.user.employeeStatus = editedUser.employeeStatus;
+    this.user.department = editedUser.department;
+    this.user.managerEmployeeCode = editedUser.managerEmployeeCode;
+    this.user.email = editedUser.email;
+    this.user.phoneNumber = editedUser.phoneNumber;
+    this.user.CL = editedUser.CL;
+    this.user.EL = editedUser.EL;
+    this.user.ML = editedUser.ML;
+    this.user.dateOfJoining = this.datepipe.transform(editedUser.dateOfJoining, "yyyy-MM-dd");
+    this.user.resignationDate = this.datepipe.transform(editedUser.resignationDate, "yyyy-MM-dd");
+    this.user.leavingDate = this.datepipe.transform(editedUser.leavingDate, "yyyy-MM-dd");
   }
   updateEmployee() {
-    this.userDataService.updateEmployeeData(this.user).subscribe((response) => {
-      this.newUserData = JSON.parse(response["_body"]).user;
-      alert('Employee Updated')
-      this.editEmpFlag = false;
-      this.user = null;
-    }, (error) => {
-      console.log(error)
-      alert(error)
-    })
+    if (localStorage.getItem('adminToken')) {
+      this.userDataService.adminupdateEmployeeData(this.user).subscribe((response) => {
+        this.newUserData = JSON.parse(response["_body"]).user;
+        alert('Employee Updated')
+        this.onloadList();
+        this.editEmpFlag = false;
+        this.user = new UserModel();
+        this.user.managerEmployeeCode = '';
+        this.user.department = '';
+        this.user.employeeStatus = '';
+        this.user.employeeType = '';
+      }, (error) => {
+        console.log(error)
+      })
+    } else {
+      this.userDataService.updateEmployeeData(this.user).subscribe((response) => {
+        this.newUserData = JSON.parse(response["_body"]).user;
+        alert('Employee Updated')
+        this.onloadList();
+        this.editEmpFlag = false;
+        this.user = new UserModel();
+        this.user.managerEmployeeCode = '';
+        this.user.department = '';
+        this.user.employeeStatus = '';
+        this.user.employeeType = '';
+      }, (error) => {
+        console.log(error)
+      })
+    }
+
   }
 
   deleteEmployee(deleteEmployee) {
-    if (confirm("Are you sure to delete " + deleteEmployee.firstName + ' ' + deleteEmployee.lastName)) {
-      this.userDataService.deleteEmployee(deleteEmployee.employeeCode).subscribe((response) => {
-        alert('Employee Deleted')
-        this.onloadList()
+    if (localStorage.getItem('adminToken')) {
+      if (confirm("Are you sure to delete " + deleteEmployee.firstName + ' ' + deleteEmployee.lastName)) {
+        this.userDataService.admindeleteEmployee(deleteEmployee.employeeCode).subscribe((response) => {
+          alert('Employee Deleted')
+          this.onloadList()
+        }, (error) => {
+          console.log(error)
+          alert(error)
+        })
+      }
+    } else {
+      if (confirm("Are you sure to delete " + deleteEmployee.firstName + ' ' + deleteEmployee.lastName)) {
+        this.userDataService.deleteEmployee(deleteEmployee.employeeCode).subscribe((response) => {
+          alert('Employee Deleted')
+          this.onloadList()
+        }, (error) => {
+          console.log(error)
+          alert(error)
+        })
+      }
+    }
+  }
+
+  loadHolidayData() {
+    if (localStorage.getItem('adminToken')) {
+      this.holidayService.adminGetHolidayList().subscribe((response) => {
+        this.holidayList = JSON.parse(response["_body"]).holidays;
+      }, (error) => {
+        console.log(error)
+        alert(error)
+      })
+    } else {
+      this.holidayService.getHolidayList().subscribe((response) => {
+        this.holidayList = JSON.parse(response["_body"]).holidays;
       }, (error) => {
         console.log(error)
         alert(error)
       })
     }
-  }
 
-  reset(resetForm) {
-    resetForm.click();
-  }
-  cancel() {
-    this.editEmpFlag = false;
-    this.user = null;
-  }
-  loadHolidayData() {
-    this.holidayService.getHolidayList().subscribe((response) => {
-      this.holidayList = JSON.parse(response["_body"]).holidays;
-    }, (error) => {
-      console.log(error)
-      alert(error)
-    })
   }
 
   addHoliday() {
-    debugger;
-    this.holidayService.addHoliday(this.holiday).subscribe((response) => {
-      this.holiday = JSON.parse(response["_body"]).holiday;
-      alert('Holiday added')
-      this.holiday.description = '';
-      this.loadHolidayData()
-    }, (error) => {
-      console.log(error)
-      alert(error)
-    })
+    if (localStorage.getItem('adminToken')) {
+      this.holidayService.adminAddHoliday(this.holiday).subscribe((response) => {
+        this.holiday = JSON.parse(response["_body"]).holiday;
+        alert('Holiday added')
+        this.holiday.description = '';
+        this.loadHolidayData()
+      }, (error) => {
+        console.log(error)
+        alert(error)
+      })
+    } else {
+      this.holidayService.addHoliday(this.holiday).subscribe((response) => {
+        this.holiday = JSON.parse(response["_body"]).holiday;
+        alert('Holiday added')
+        this.holiday.description = '';
+        this.loadHolidayData()
+      }, (error) => {
+        console.log(error)
+        alert(error)
+      })
+    }
+
   }
 
   editHoliday(holiday) {
@@ -293,26 +374,52 @@ export class DataEntryComponent implements OnInit {
   }
 
   updateHoliday() {
-    this.holidayService.updateHoliday(this.holiday).subscribe((response) => {
-      this.holiday = JSON.parse(response["_body"]).holiday;
-      alert('Holiday Updated')
-      this.editHolidayFlag = false;
-      this.loadHolidayData()
-    }, (error) => {
-      console.log(error)
-      alert(error)
-    })
-  }
-
-  deleteHoliday(holiday) {
-    if (confirm("Are you sure to delete " + this.datepipe.transform(holiday.date, "dd-MM-yyyy"))) {
-      this.holidayService.deleteholiday(holiday.date).subscribe((response) => {
-        alert('Employee Deleted')
+    if (localStorage.getItem('adminToken')) {
+      this.holidayService.adminUpdateHoliday(this.holiday).subscribe((response) => {
+        this.holiday = JSON.parse(response["_body"]).holiday;
+        alert('Holiday Updated')
+        this.editHolidayFlag = false;
+        this.holiday = new HolidayModel();
         this.loadHolidayData()
       }, (error) => {
         console.log(error)
         alert(error)
       })
+    } else {
+      this.holidayService.updateHoliday(this.holiday).subscribe((response) => {
+        this.holiday = JSON.parse(response["_body"]).holiday;
+        alert('Holiday Updated')
+        this.editHolidayFlag = false;
+        this.holiday = new HolidayModel();
+        this.loadHolidayData()
+      }, (error) => {
+        console.log(error)
+        alert(error)
+      })
+    }
+  }
+
+  deleteHoliday(holiday) {
+    if (localStorage.getItem('adminToken')) {
+      if (confirm("Are you sure to delete " + this.datepipe.transform(holiday.date, "dd-MM-yyyy"))) {
+        this.holidayService.admindeleteholiday(holiday.date).subscribe((response) => {
+          alert('Employee Deleted')
+          this.loadHolidayData()
+        }, (error) => {
+          console.log(error)
+          alert(error)
+        })
+      }
+    } else {
+      if (confirm("Are you sure to delete " + this.datepipe.transform(holiday.date, "dd-MM-yyyy"))) {
+        this.holidayService.deleteholiday(holiday.date).subscribe((response) => {
+          alert('Employee Deleted')
+          this.loadHolidayData()
+        }, (error) => {
+          console.log(error)
+          alert(error)
+        })
+      }
     }
   }
 }
